feat: unwrap MCP namespace to flat exports + barrel for tree-shaking
Convert `export namespace MCP { ... }` (930 lines) to flat named exports
with `export * as MCP` barrel. Third leg of the cli/error.ts import chain —
this pulled in @modelcontextprotocol/sdk (3 transports) and open (browser
launcher) just for MCP.Failed.isInstance() checks.

33 tests pass, full typecheck clean.
